Allow comments anywhere in the document
Despite how loose the INI spec is, some languages allow comments at the end of
the line and not having the ability to highlight those may cause confusion to
the people who are using parsers that allow that.

This piggy backs of the work by @j2ghz in #11 with an additional line for the
comments starting with '#'.

Fixes #26, #10 and #11.