Create a separate grammar for `.desktop` files
As we are well aware, every tool has their own intereptation of the ini
syntax and up until now, I've dumped them all into a single grammar. I
learnt the other day you can split these out and restrict it to certain
file types so it makes sense to so that we can treat `.desktop` comments
differently to others.

Fixes #33